fixed config option
git-svn-id: svn+ssh://svn.lighttpd.net/lighttpd/branches/lighttpd-merge-1.4.x@897 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/mod_evasive.c
@@ -26,7 +26,7 @@
  */
 
 typedef struct {
-	int max_conns;
+	unsigned short max_conns;
 } plugin_config;
 
 typedef struct {
